requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.2.1
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1.7
django-debug-toolbar==5.1.0
executing==2.2.0
ipdb==0.13.13
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-decouple==3.8
sqlparse==0.5.3
stack-data==0.6.3
static3==0.7.0
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
